ATAS Trading Bot: Demand-Index-Indikator gegen echten Build korrigiert
Erster Build-Versuch bei Mike (19.09.2026) lieferte zwei Fehler: OFT.Attributes
braucht eine eigene DLL-Referenz (nicht in ATAS.Indicators.dll enthalten), und
candle.Volume ist decimal statt double. Code auf decimal umgestellt, keine
Casts mehr nötig.

+++ 02 Projekte/ATAS Trading Bot (NQ-Abpraller-Setup)/code/NqDemandIndex.cs
@@ -1,13 +1,11 @@
 // Phase 0/1 — erster Baustein des ATAS Trading Bots, siehe Projekt-Notiz im übergeordneten Ordner.
 //
-// WICHTIG: Diese Datei ist gegen die öffentliche ATAS-Doku geschrieben, nicht gegen eine echte
-// SDK-Installation getestet (dieser Cloud-Session steht kein Windows/ATAS zur Verfügung). Beim
-// ersten Build sind Fehler an folgenden Stellen am wahrscheinlichsten:
-//   - Namespace/Attribute für [DisplayName]/[Parameter] können in deiner SDK-Version anders heißen
-//   - Der genaue Konstruktor/die Indexer-API von ValueDataSeries kann leicht abweichen
-//   - candle.Volume könnte decimal statt double sein (Cast unten ggf. anpassen)
-// Schick mir bei einem Build-Fehler einfach die genaue Fehlermeldung, dann korrigiere ich gezielt
-// statt hier weiter zu raten.
+// Stand 19.09.2026: gegen Mikes echte ATAS-Installation gebaut und korrigiert (erster Build-Versuch).
+// Zwei Korrekturen gegenüber der ersten Fassung:
+//   - using OFT.Attributes braucht eine eigene Referenz auf OFT.Attributes.dll (im ATAS-Platform-
+//     Ordner, nicht in ATAS.Indicators.dll enthalten)
+//   - candle.Volume ist decimal, nicht double — komplette Rechnung läuft jetzt in decimal, keine
+//     Casts mehr nötig
 //
 // Formel: [[RG-Trading Indikator - Demand Index]] (Sibbet), referenziert in
 // [[NQ Abpraller-Setup Checkliste]] Punkt 4 als letzter Bestätigungsfilter.
@@ -22,7 +20,7 @@ namespace RgTrading.Indicators
     {
         private readonly ValueDataSeries _demandIndex = new ValueDataSeries("Demand Index");
         private readonly ValueDataSeries _zeroLine = new ValueDataSeries("Nulllinie");
-        private double _cumulative;
+        private decimal _cumulative;
 
         public NqDemandIndex() : base(true)
         {
@@ -44,10 +42,10 @@ namespace RgTrading.Indicators
 
             // Schritt 1+2: relative Preisänderung = (Close - Open) / Open
             var openPrice = candle.Open == 0 ? 1 : candle.Open; // Division durch 0 vermeiden
-            var relativeChange = (double)((candle.Close - candle.Open) / openPrice);
+            var relativeChange = (candle.Close - candle.Open) / openPrice;
 
             // Schritt 3+4: Volumenkomponente kumulieren
-            var volumeComponent = (double)candle.Volume * relativeChange;
+            var volumeComponent = candle.Volume * relativeChange;
             _cumulative += volumeComponent;
 
             _demandIndex[bar] = _cumulative;
